Task 5: Implement isSorted() Member Function and Improve search() API
isSorted() member function which will return true if the List is currently sorted, and false if it is not. This function makes the overload search more effiecient to use. The search() function takes a string to search for in the list, and it returns the index where the value is found in the list.